move tests for settings to dedicated file
