+++ Makefile
@@ -5,7 +5,7 @@ test:
 # run tests on any change
 continuous_test:
 #	ptw *.py
-	ptw main.py test_*.py
+	ptw main.py settings.py test_*.py
 
 
 # generate coverage report as html
