+++ test_main.py
@@ -187,15 +187,3 @@ def test_check_metrics_endpoint(client: TestClient) -> None:
         'text/plain; version=0.0.4')
     assert response.text.startswith(
         "# HELP python_gc_objects_collected_total Objects collected during gc")
-
-
-def test_settins_init():
-    settings = Settings()
-    assert settings != None
-
-
-def test_settins_get_db_url():
-    settings = Settings()
-    assert settings.db_url != None
-    assert isinstance(settings.db_url, str) is True
-    assert len(settings.db_url) > 5
